Prevent randomized button colors from being too dark
The purpose of this change is to ensure the font inside button icons is always readable, notably when used as a note or in the pause menu. It does mean some of the darker colors will never be selected, but I think it's more important to keep the game accessible and playable no matter what.