Rewrite Deku backroom GS to be more readable.
I rewrote some of the here()s to just be a room instead, and sorted it all out to be less of a mess.

This should be the same logically, just actually human-readable.

I've drafted it because I'm not sure that the trick for burning the vertical webs is actually necessary. 2/3 fire requirements are absolutely trivial with bow (one is literally the exact same as the intended method for torches room in water temple, and the other is even easier), and the only slightly awkward one is in the main b1 room (you need to stand to the right of the torch and shoot. It doesn't look like the fire will go through because you're to the side of the torch, but it does because OoT).