Restore Original Cosmetic Model/Icon Colors
Currently, if a patch file was generated with a non-default heart color, non-default tunic colors, or with "Item Model Color Matches Cosmetics", the models and icons of these cosmetic changes are not overwritten back to the default when patching with cosmetics that use default heart/magic/tunic colors or without "Item Colors Match Cosmetics" enabled. This commit should restore the original colors from the base ROM in those cases.

Fixes #1075.