Add new '_dungeon' and '_overworld' variants of WotH and Barren hints.
Adds 'woth_dungeon', 'woth_overworld', 'barren_dungeon', and 'barren_overworld' as hint types. These are the same as their base types but restrict the location/area to a dungeon or an overworld region respectively. In the event that the dungeon hints run out of dungeons for one reason or another, those will fall back to allowing an overworld location. The overworld variants do not have a fallback if there are no valid overworld locations, and that hint would be skipped instead.