Update Tournament hint distribution.
Removed Skull Mask from ever being considered an always hint without the hint distribution specifically requesting it. Tournament hints no longer group the Temple of Time gossip stones into a single hint and now include one barren dungeon hint (falls back to an overworld region if none are available) and a generic barren hint (which would be an overworld region as the barren dungeon limit is 1). Updated the tooltip of the Tournament hint distribution to be accurate.