replace ow/dun hint balance with auto balancer
The overworld/dungeon options for woth/barren were removed because they lead to a situation where the player can frequently determine a large amount of meta knowledge about the seed if they recognize that a particular type failed to generate. Instead this now will try auto balance between overworld and dungeon hints for barren. In the future, it might make sense to add hint distribution parameters that specify the weight.