Fix the logic issue with Crater Upper to Central in ER
Any ER setting will now disable the fairy use here (which is only relevant in quad dmg or OHKO). This is more restricted than it has to be, but I don't think it's worth taking the risk to have a more "accurate" condition which will have to be constantly updated.