Minor fix for DC lobby wall
This is logically equivalent, but more closely matches what's happening in the game: There's an event (breaking the wall in front of the lobby) which can be done with explosives or hammer (`can_blast_or_smash`) or with the bomb flowers (`Progressive_Strength_Upgrade`). MQ already has the rule like this.