Default Presets Update (DDR Weekly & Multiworld added, Hell Mode update)
Adds the DDR Weekly and Multiworld presets. These both exist on the web version already, so this is to bring the release builds into sync with the website's defaults. DDR Weekly settings were provided by shiroaeli. The Multiworld preset is based on the old Multiworld Tournament settings with 'Skip Some Minigame Phases' additionally enabled as that is one of the few new settings that I am confident would be wanted.

Additionally removes the excluded locations from the Hell Mode preset as they no longer make any sense to be there. This should not conflict with #1190 which is also necessary to fully update Hell Mode.