Avoid failing to place hints on fixed hint types with multiple copies.
This particularly affects uses of the new tournament hint distribution on
settings where the trial+always hints use up an even number of gossip
stones and leave a small enough odd number of stones (the total number
of available stones is odd in this dist).

The S4 settings preset is not affected as the number of always locations,
predetermined by the preset, leaves enough space for the fixed
WoTH + barren hints that have multiple copies.