Minor Medallion Reordering
This PR feels rather dumb.... noticed that the PR to make the spoiler log locations be in a more sensible order meant that the order of Bongo and Twinrova's boss rewards got swapped. Screws with the website display of dungeon prizes which is how I noticed it, and generally Shadow is listed before Spirit anyway.