Clarify IGC hints and +1 junk
Updates "Ganon's Castle is barren/WotH" hints to specify that it's *inside* Ganon's Castle. There'd been a bit of confusion around it in the past about whether "Ganon's Castle" was the interior or exterior or both, and this should eliminate that.

Also adds a junk hint saying the player is good, because having one that only said they were bad with no counterpart felt kinda mean.