Fix same hint area checks in ER to be compatible with random warp songs
This fixes an ER issue where a randomized warp song leading to inside Kak Potion Shop or Impas House could prevent "same hint area" checks from working properly.
In other words, the entrances to Kak Potion Shop (or Impas House in cow shuffle) could occasionally end up in different areas, and we want to avoid that so there is no ambiguity as to which entrance determines the hint.