Fix error when file extension contains capital Z.
When attempting to generate a randomized rom, if the file extension contains a capital Z, the program will throw an error saying that the rom is invalid. Adding the .lower() fixes this issue. 
This may be a MacOS/Linux specific issue, since Windows usually does not care about letter cases.