Remove the fake key from water temple
This key only exists in logic. It's not real, and doesn't appear in game.

There's no real reason to keep it around, since it's functionally *identical* to just having the key counts in water be reduced by one. The fire one should still stay because that's actually tied to settings, but the only setting that took this fake key out of logic was turning MQ on...which already uses a separate set of logic files.

It's also pointlessly confusing for anyone reading the logic files, since they'll end up assuming you need one more key than rando says you have to, and I don't think anyone wants to keep explaining what a "fake logical key" is.