Correctly check plando'ed one-way priorities.
The algorithm was checking two orders deep:
- Serenade of Water Warp -> Lake Hylia pointed to Bolero of Fire Warp ->
  DMC Central Local.
- Bolero of Fire Warp -> DMC Central Local pointed to Lost Woods -> Zora
  River.
- Lost Woods -> Zora River pointed to nothing, of course.

We want to look at the regions where we've connected our one-ways, to
determine if we still need to place entrances to the priority regions.
To do this, we look at the region Serenade of Water Warp -> Lake
Hylia now points to (DMC Central Local).

We still need the list of entrances *replaced* so that we can remove
them from consideration as replacement candidates, but we can't assume
that they themselves have replacements (or that
entrance.replaces.replaces == entrance, or that they always replace
warp song destinations...).

Fixes #1300.